requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
Pillow==10.1.0
